correct deleted images

+++ controllers/profileController.js
@@ -3,6 +3,7 @@ const User = require('../models/User');
 var mongoose = require('mongoose');
 const path = require('path');
 const fs = require('fs');
+const { error } = require('console');
 
 exports.getProfile = (req, res) => {
     Template.find({ owner: req.auth.userId, favorite: true })
@@ -90,49 +91,39 @@ exports.getOtherProfile = (req, res) => {
 }
 
 exports.putProfile = (req, res) => {
-    let toChange = {};
-    let old_banner = "";
-    let old_profile = "";
-    if (req.files.profileImage){
-        toChange["profile_image"] = req.files.profileImage[0].filename;
-        old_profile = User.findOne({_id:req.auth.userId})
-            .then(u => {if (u.profile_image != "d_profile.png"){
-                return u.profile_image
-            }})
-    }
-    if (req.files.bannerImage){
-        toChange["banner_image"] = req.files.bannerImage[0].filename;
-        old_banner = User.findOne({_id:req.auth.userId})
-            .then(u => {if (u.banner_image != "d_banner.png"){
-                return u.banner_image
-            }})
-    }
-    if (req.body.username){
-        toChange["username"] = req.body.username;
-    }
-    const id = {_id: req.auth.userId};
-    User.findOneAndUpdate(id,toChange)
-        .then(() => {
-            console.log("0");
-            if (old_banner != ""){
-                console.log("0.5");
-                console.log(old_banner);
-                console.log(path.join(path.dirname(__dirname), "images","test"));
-                fs.unlink(path.join(path.dirname(__dirname), 'images',old_banner), (err) => {
-                    console.log("1");
-                    if (err) {
-                      console.error('Erreur lors de la suppression du fichier :', err);
-                      return;
+    const id = {_id:req.auth.userId};
+    User.findOne(id)
+        .then( user => {
+            const old_banner = user.banner_image;
+            const old_profile = user.profile_image;
+            let toChange = {};
+            if (req.files.profileImage){
+                toChange["profile_image"] = req.files.profileImage[0].filename;
+            }if (req.files.bannerImage){
+                toChange["banner_image"] = req.files.bannerImage[0].filename;
+            }if (req.body.username){
+                toChange["username"] = req.body.username;
+            }
+            User.findOneAndUpdate(id,toChange)
+                .then(() => {
+                    let suppError = {};
+                    if (req.files.profileImage && old_banner != "d_banner.png"){
+                        fs.unlink(path.join(path.dirname(__dirname), 'images',old_banner), (err) => {
+                            if (err) {
+                                suppError["banner"] = "Erreur lors de la suppression du fichier :", err;
+                            }});
+                    }if (req.files.bannerImage && old_profile != "d_profile.png"){
+                        fs.unlink(path.join(path.dirname(__dirname), 'images',old_profile), (err) => {
+                            if (err) {
+                                suppError["profile"] = "Erreur lors de la suppression du fichier :", err;
+                            }});
                     }
-                    console.log('Fichier supprimé avec succès :', old_banner);
-                  });
-            }
-            console.log("2");
-            if (old_profile){
-                fs.unlink(old_profile);
-            }
-            res.status(200).json({message: "Username changed" });
-        })
-        .catch(error => res.status(402).json({ error , message: "Can't change the username"}))
+                    if (Object.keys(suppError).length > 0){
+                        res.status(401).json({suppError, message:"Can't delete the file"});
+                        return;
+                    }
+                    res.status(200).json({message:"Profile updated"});
+                })
+                .catch(error => res.status(401).json({error, message:"Can't update the profile"}))
+        }).catch(error => res.status(401).json({error, message:"Can't find the user"}));
 }
-
